requirements.txt
requests
colorama
discord.py-self
gtts
pyfiglet
python-whois
python-dateutil